# faster cbor serialisation/deserialisation