--- a/blockchain/src/genesis.rs
+++ b/blockchain/src/genesis.rs
@@ -54,7 +54,7 @@ impl raw_cbor::de::Deserialize for Body {
             let t = raw.cbor_type()?;
             if t == raw_cbor::Type::Special {
                 let special = raw.special()?;
-                assert_eq!(special, raw_cbor::de::Special::Break);
+                assert_eq!(special, raw_cbor::Special::Break);
                 false
             } else {
                 slot_leaders.push_back(raw_cbor::de::Deserialize::deserialize(raw)?);