--- a/blockchain/src/lib.rs
+++ b/blockchain/src/lib.rs
@@ -11,6 +11,8 @@ extern crate test;
 extern crate serde_derive;
 extern crate serde;
 
+extern crate raw_cbor;
+
 pub mod types;
 pub mod genesis; /* genesis block related value */
 pub mod normal; /* normal block related value */